backend/requirements.txt
blinker==1.6.3
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
flask==3.0.0
Flask-Cors==4.0.0
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
MarkupSafe==2.1.3
numpy==1.26.1
python-dotenv==1.0.0
requests==2.31.0
scikit-learn==1.2.2
scipy==1.11.3
threadpoolctl==3.2.0
urllib3==2.0.7
werkzeug==3.0.0
zipp==3.17.0
